Add jit merge point.
